+++ internal/awslambda/extension/api/doc.go
@@ -1,0 +1,8 @@
+// Package api defines the interaction that is done between
+// the aditional runtime (the current running extension)
+// and the main business logic function running along aside it.
+//
+// For further infomation regarding lambda extensions:
+//
+//	https://docs.aws.amazon.com/lambda/latest/dg/runtimes-extensions-api.html#runtimes-extensions-registration-api
+package api